Slight refactor to split out config and clock
